Move brightness logic to Pixels

+++ src/pixels.h
@@ -11,7 +11,7 @@ namespace Pixels
 
   void setChannels();
   uint32_t allChannels(uint8_t);
-  void setGlobalBrightness(uint8_t);
+  void setGlobalBrightness(uint16_t);
   void fade(CRGB, uint8_t, uint8_t);
   void clear(CRGB, uint8_t, uint8_t);
   void ebbAndFlow(CRGB[], uint8_t[], uint8_t);
